requirements.txt
elevenlabs>=1.6.0
python-dotenv>=1.0.0
